Moved term registry to `terms.registry` submodule
